Save filename for uploaded document
This is necessary because the file extension alone
is not identifiable without a doubt
(take my.file.txt vs. my.file.tar.gz, for example).
By storing the whole file name in the file and the
database, we can find the file again later that
belongs to the document entry.